+++ src/frontend/org/voltdb/expressions/ParameterValueExpression.java
@@ -183,10 +183,12 @@ public class ParameterValueExpression extends AbstractValueExpression {
         if (m_valueType != null && m_valueType != VoltType.NUMERIC) {
             return;
         }
-        VoltType fallbackType = VoltType.FLOAT;
+
+        // Typically BIGINT or DECIMAL as the exact number constants.
+        VoltType fallbackType = VoltType.DECIMAL;
         if (m_originalValue != null) {
             m_originalValue.refineOperandType(VoltType.BIGINT);
-            fallbackType = m_originalValue.getValueType(); // Typically BIGINT or FLOAT.
+            fallbackType = m_originalValue.getValueType();
         }
         m_valueType = fallbackType;
         m_valueSize = m_valueType.getLengthInBytesForFixedTypes();
